requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
sse-starlette>=1.6.5

# AI Integration (只用核心SDK，不用langchain简化依赖)
openai>=1.12.0
anthropic>=0.18.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
markdown>=3.5.2

# Report Generation
jinja2>=3.1.3
reportlab>=4.0.0
xlsxwriter>=3.1.0

# Test Data Generation
faker>=22.0.0

# Utilities
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
